+++ src/log.ts
@@ -10,15 +10,40 @@
  * and keeping track of statistics
  */
 
+import * as util from 'util';
+import * as stream from 'stream';
 /* eslint-disable @typescript-eslint/no-var-requires */
 const chalk = require('chalk');
 
-export const logConfig = {
+export interface LogConfig {
+	/** Whether debug level log (starting with [D]) should be printed */
+	printDebug?: boolean;
+	/** Whether verbose logs (starting with [D]) should be printed */
+	printVerbose?: boolean;
+	/**
+	 * The writable stream for logs. Defaults to stdout. To log to a file, use
+	 * ```js
+	 * fs.createWriteStream(__dirname + '/mwn.log', {
+	 *     flags: 'a',
+	 *     encoding: 'utf8'
+	 * })
+	 * ```
+	 */
+	stream?: stream.Writable;
+}
+
+const logConfig: LogConfig = {
 	printDebug: true,
 	printVerbose: true,
+	stream: process.stdout as stream.Writable,
 };
 
-export function updateLoggingConfig(options: Partial<typeof logConfig>): void {
+/**
+ * Configure global logging options.
+ *
+ * Note: To suppress API warnings, use `suppressAPIWarnings` flag in bot instance options instead.
+ */
+export function updateLoggingConfig(options: LogConfig): void {
 	Object.assign(logConfig, options);
 }
 
@@ -28,7 +53,7 @@ type item = any;
 /**
  * Custom Logging function
  *
- * Writes Logs to console, stringifies objects first
+ * Writes logs to console or file, stringifies objects first
  *
  * @param obj
  */
@@ -56,7 +81,7 @@ export function message(msg: item) {
 	msg = colorize(msg);
 
 	if (msg.trim && msg.trim().length > 0) {
-		msg = chalk.gray('[' + humanDate() + '] ') + msg;
+		msg = colorize('[' + humanDate() + '] ', 'gray') + msg;
 	}
 
 	if (!logConfig.printVerbose && msg.indexOf('[V]') >= 0) {
@@ -64,7 +89,7 @@ export function message(msg: item) {
 	} else if (!logConfig.printDebug && msg.indexOf('[D]') >= 0) {
 		// Supressing output of debug message
 	} else {
-		console.log(msg);
+		writeToLog(msg);
 	}
 }
 
@@ -75,7 +100,7 @@ export function message(msg: item) {
 export function debug(obj: item) {
 	// Print indented JSON
 	let msg = JSON.stringify(obj, null, 4);
-	console.log(chalk.gray(msg));
+	writeToLog(colorize(msg, 'gray'));
 }
 
 /**
@@ -83,7 +108,7 @@ export function debug(obj: item) {
  * @param obj
  */
 export function error(obj: Error) {
-	console.error(chalk.red('[E] ' + obj.message));
+	writeToLog(colorize('[E] ' + obj.message, 'red'));
 	let stringified;
 	try {
 		stringified = JSON.stringify(obj, null, 4);
@@ -91,9 +116,9 @@ export function error(obj: Error) {
 		// Circular object?
 		stringified = `Failed to stringify error: ${e.message}`;
 	}
-	console.log(chalk.gray(stringified));
+	writeToLog(colorize(stringified, 'gray'));
 	if (obj.stack) {
-		console.log(chalk.gray(obj.stack));
+		writeToLog(colorize(obj.stack, 'gray'));
 	}
 }
 
@@ -119,9 +144,16 @@ let colorMap = {
  * Colors the messages by searching for specific indicator strings
  *
  * @param {string} msg
+ * @param {string} colorName
  * @returns {string}
  */
-export function colorize(msg: item) {
+export function colorize(msg: item, colorName?: string) {
+	if (logConfig.stream !== process.stdout) {
+		return msg;
+	}
+	if (colorName) {
+		return chalk[colorName](msg);
+	}
 	for (let [code, color] of Object.entries(colorMap)) {
 		if (msg && msg.indexOf && msg.startsWith(code)) {
 			return chalk[color](msg);
@@ -129,6 +161,14 @@ export function colorize(msg: item) {
 	}
 
 	return msg;
+}
+
+function writeToLog(message: string) {
+	if (logConfig.stream === process.stdout) {
+		console.log(message);
+	} else {
+		logConfig.stream.write(util.format.apply(null, [message]) + '\n');
+	}
 }
 
 /**